// src/types/utools.d.ts
export interface DBItem<T> {
  _id: string
  _rev?: string
  data: T
}

export interface DBRes<T> {
  _id: string
  ok: boolean
  data: T
  _rev?: string
  error?: any
}